Crate spl_generic_token

Modules

associated_token_account

Partial SPL Associated Token Account declarations to avoid a dependency on the spl-associated-token-account crate.

generic_token

Minimum viable SPL Token parsers to avoid a dependency on the spl-token and spl-token-2022 crates. Users may use the generic traits directly, but this requires them to select the correct implementation based on the account’s program id. generic_token::Account and generic_token::Mint abstract over this and require no knowledge of the different token programs on the part of the caller at all.

token

Partial SPL Token declarations to avoid a dependency on the spl-token crate.

token_2022

Partial SPL Token declarations to avoid a dependency on the spl-token-2022 crate.

Functions

is_known_spl_token_id

Check if the provided program id as a known SPL Token program id

spl_token_ids

Returns all known SPL Token program ids